+++ BloodCraftUI/UI/CustomLib/Cells/IFormedCell.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using BloodCraftUI.UI.UniverseLib.UI.Widgets.ScrollView;
+
+namespace BloodCraftUI.UI.CustomLib.Cells;
+
+public interface IFormedCell : ICell
+{
+    public int CurrentDataIndex { get; set; }
+    public Action<int> OnClick { get; set; }
+}